requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.28.3
botocore==1.31.3
bs4==0.0.1
build==0.10.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
cryptography==41.0.2
cssselect==1.2.0
dataclasses-json==0.5.9
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
dnspython==2.4.0
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.7.1
et-xmlfile==1.1.0
fake-useragent==1.1.3
filetype==1.2.0
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
jmespath==1.0.1
joblib==1.3.1
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
langchain==0.0.233
langsmith==0.0.5
loguru==0.7.0
lxml==4.9.3
Markdown==3.4.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
olefile==0.46
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parse==1.19.1
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==10.0.0
pinecone-client==2.2.2
pycparser==2.21
pydantic==1.10.11
pyee==8.2.2
PyJWT==2.7.0
pypandoc==1.11
pyppeteer==1.0.2
pyproject_hooks==1.0.0
pyquery==2.0.0
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
requests-html==0.10.0
rpds-py==0.8.10
s3transfer==0.6.1
singlestoredb==0.8.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
sqlparams==5.1.0
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.8.1
uritemplate==4.1.1
urllib3==1.26.16
w3lib==2.1.1
websockets==10.4
win32-setctime==1.1.0
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.16.2
singlestoredb==0.8.1